Kyc documents upload fix and flow rerrangement
